Learned from a template
Basically, 2 pointers, 1 counter equal to the length of the target, 1 target dictionary whose values are the number of their corresponding key occurrence, and 1 stored substring length initially set to more than the max of both strings or infinity(since we're optimizing the minimum). 

The latter pointer moves through the source string, and when the current element can be found in dictionary, the value in the dictionary is reduced and so is the counter, because one less element in the target needs to be found. If the current element can't be found in dictionary, only the latter pointer will increase. If redundant elements are found, which means the needed quantity of target is 0 already, the quantity will decrease to negative which indicates there's redundancy, and the counter would not change. 

After moving the latter pointer, the function checks if counter is 0. If counter is not equal to 0, the substring remains unchanged and goes to the new iteration. Otherwise, it means the substring has included all the elements in the target, and the length of the substring would be compared to the stored length, and furthermore the function will choose the shortest one. 

After finding a valid substring, the function needs to continue exploring whether the substring is optimal(shortest), and thus it will check the front pointer and move it. For every iteration the element encountered would increase its number by 1 in the target dictionary(if it is in the dictionary). If the element it points to is found in dictionary and it's quantity is 0, counter will increase by 1, which means the function needs to find 1 of this element with the front pointer in the future.